put ListOfProduct to ml_sumcheck::data_structures
